pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "packrecorder"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "PySide6>=6.6.0",
  "opencv-python>=4.9.0",
  "pyzbar>=0.1.9",
  "numpy>=1.26.0",
  "pyserial>=3.5",
  "pillow>=10.0.0",
  "hidapi>=0.14.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-qt>=4.4"]
build = ["pyinstaller>=6.0"]
admin = ["fastapi>=0.110", "uvicorn>=0.27"]
background = ["psutil>=5.9.0"]

[project.scripts]
packrecorder = "packrecorder.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
